refactor: move testFolder to a general location
