+++ core/aws-lsp-core/src/index.ts
@@ -14,3 +14,4 @@ export * as textUtils from './util/text'
 export * as timeoutUtils from './util/timeoutUtils'
 export * from './util/awsError'
 export * from './base/index'
+export * from './test/testFolder'
