+++ server/aws-lsp-codewhisperer/src/language-server/agenticChat/tools/toolShared.ts
@@ -1,5 +1,3 @@
-import * as path from 'path'
-
 export const maxToolResponseSize = 30720 // 30KB
 
 export enum OutputKind {
@@ -12,17 +10,4 @@ export interface InvokeOutput {
         kind: OutputKind
         content: string
     }
-}
-
-export function sanitizePath(inputPath: string, homePath: string): string {
-    let sanitized = inputPath.trim()
-
-    if (sanitized.startsWith('~')) {
-        sanitized = path.join(homePath, sanitized.slice(1))
-    }
-
-    if (!path.isAbsolute(sanitized)) {
-        sanitized = path.resolve(sanitized)
-    }
-    return sanitized
 }
